chore: separate adapters and providers folders
